+++ CHANGELOG.md
@@ -1,10 +1,18 @@
+## 1.0.9
+
+- Added showPanButtons and showResetButton parameters to the constructor (both default false)
+- Changed zoomButtonPosition and zoomButtonColor to buttonPosition and buttonColor (breaking)
+- Clamped zoomLevel, zoomCenter and panTo inputs to acceptable values
+- Added an extra value in the onTap callback function. Now (imageOffset, windowOffset) (breaking)
+- Code optimization and bug fixes
+
 ## 1.0.8
 
 - Added parameter 'interactive' (default = true). If set to false, you can/should add your own keyboard/mouse/gesture detector on top of 
   the zoomify widget, or use the controller to interact with the image.
 - Added parameter 'pantTo: Offset' to controller.(animate)PanAndZoom function. Moves the given point, relative to the original max image, 
   to the center of the window
-- Added parameter 'fitImage' (default = true). If set to fals, allows the image to become smaller then the winde
+- Added parameter 'fitImage' (default = true). If set to false, allows the image to become smaller then the winde
 
 ## 1.0.7